Implement a GetEntry trait
