+++ fluent/src/resolve.rs
@@ -11,7 +11,8 @@ use std::collections::hash_map::DefaultHasher;
 use std::collections::HashMap;
 use std::hash::{Hash, Hasher};
 
-use super::context::{Entry, MessageContext};
+use super::context::MessageContext;
+use super::entry::GetEntry;
 use super::types::FluentValue;
 use fluent_syntax::ast;
 
@@ -132,23 +133,13 @@ impl ResolveValue for ast::Expression {
             ast::Expression::MessageReference { ref id } if id.name.starts_with('-') => env
                 .ctx
                 .entries
-                .get(&id.name)
-                .and_then(|entry| match entry {
-                    Entry::Message(_) => None,
-                    Entry::Term(term) => Some(term),
-                    Entry::Function(_) => None,
-                })
+                .get_term(&id.name)
                 .ok_or(FluentError::None)?
                 .to_value(env),
             ast::Expression::MessageReference { ref id } => env
                 .ctx
                 .entries
-                .get(&id.name)
-                .and_then(|entry| match entry {
-                    Entry::Message(message) => Some(message),
-                    Entry::Term(_) => None,
-                    Entry::Function(_) => None,
-                })
+                .get_message(&id.name)
                 .ok_or(FluentError::None)?
                 .to_value(env),
             ast::Expression::ExternalArgument { ref id } => env
@@ -198,24 +189,14 @@ impl ResolveValue for ast::Expression {
                 let attributes = if id.name.starts_with('-') {
                     env.ctx
                         .entries
-                        .get(&id.name)
-                        .and_then(|entry| match entry {
-                            Entry::Message(_) => None,
-                            Entry::Term(term) => Some(term),
-                            Entry::Function(_) => None,
-                        })
+                        .get_term(&id.name)
                         .ok_or(FluentError::None)?
                         .attributes
                         .as_ref()
                 } else {
                     env.ctx
                         .entries
-                        .get(&id.name)
-                        .and_then(|entry| match entry {
-                            Entry::Message(message) => Some(message),
-                            Entry::Term(_) => None,
-                            Entry::Function(_) => None,
-                        })
+                        .get_message(&id.name)
                         .ok_or(FluentError::None)?
                         .attributes
                         .as_ref()
@@ -230,16 +211,7 @@ impl ResolveValue for ast::Expression {
                 Err(FluentError::None)
             }
             ast::Expression::VariantExpression { id, key } if id.name.starts_with('-') => {
-                let term = env
-                    .ctx
-                    .entries
-                    .get(&id.name)
-                    .and_then(|entry| match entry {
-                        Entry::Message(_) => None,
-                        Entry::Term(term) => Some(term),
-                        Entry::Function(_) => None,
-                    })
-                    .ok_or(FluentError::None)?;
+                let term = env.ctx.entries.get_term(&id.name).ok_or(FluentError::None)?;
                 let variants = match term.value.elements.as_slice() {
                     [ast::PatternElement::Placeable(ast::Expression::SelectExpression {
                         expression: None,
@@ -290,12 +262,7 @@ impl ResolveValue for ast::Expression {
 
                 env.ctx
                     .entries
-                    .get(&callee.name)
-                    .and_then(|entry| match entry {
-                        Entry::Message(_) => None,
-                        Entry::Term(_) => None,
-                        Entry::Function(func) => Some(func),
-                    })
+                    .get_function(&callee.name)
                     .ok_or(FluentError::None)
                     .and_then(|func| {
                         func(resolved_unnamed_args.as_slice(), &resolved_named_args)
